Realice varios cambios
1- Realice ruta get/post BILLS
2- Realice ruta get/post PARTNERS
3- cambie los CREATE por findOrCreate
4- BACK funcionando perfecto
5- Cambie Solo FORCE A TRUE

+++ src/models/Partners.js
@@ -39,7 +39,7 @@ module.exports = (sequelize) => {
         allowNull: false,
       },
       mobile: {
-        type: DataTypes.INTEGER,
+        type: DataTypes.BIGINT, // LE PUSE BIGINT porque de esa forma almacena mas numeros
       },
     },
     {
